Useless twice double conversion

+++ src/lecture_csv.c
@@ -235,7 +235,7 @@ void completer_mat_duels(t_mat_int_dyn* duels, int nb_electeurs) {
         for (int candidat2_id = 0; candidat2_id < candidat1_id; candidat2_id++) {
             int pourcentage_voix =
                 ((double) duels->elems[candidat1_id][candidat2_id])
-                / ((double)nb_electeurs) * 100;
+                / nb_electeurs * 100;
 
             duels->elems[candidat1_id][candidat2_id] = pourcentage_voix;
             duels->elems[candidat2_id][candidat1_id] = 100 - pourcentage_voix;
